Begin fix desktop

+++ composeApp/src/androidMain/kotlin/org/trevor/pcup/Platform.android.kt
@@ -10,14 +10,14 @@ import co.touchlab.kermit.Logger
 
 class AndroidPlatform(private val ctx: Context) : Platform {
     override val name: String = "Android ${Build.VERSION.SDK_INT}"
-    override val battery: Float?
+    override val battery: Int?
         get() {
             val battery = ctx.getSystemService(BATTERY_SERVICE) as? BatteryManager
             if (battery == null) {
                 Logger.i("could not get BATTERY_SERVICE")
                 return null
             }
-            return battery.getIntProperty(BatteryManager.BATTERY_PROPERTY_CAPACITY).toFloat()
+            return battery.getIntProperty(BatteryManager.BATTERY_PROPERTY_CAPACITY)
         }
 
     init {
